Implemented suggested changes
